refactor: update code accordindg to comments
