+++ App/commonStyles/textStyles.js
@@ -1,0 +1,5 @@
+export const commonTextStyle = {
+  style: {
+    fontFamily: 'Nunito-Light',
+  },
+};
