--- App/consts/globalProps.js
@@ -1,11 +1,0 @@
-export const customTextInputProps = {
-  style: {
-    fontFamily: 'Nunito-Light',
-  },
-};
-
-export const customTextProps = {
-  style: {
-    fontFamily: 'Nunito-Light',
-  },
-};
