--- App/utils/utils.js
@@ -1,6 +1,0 @@
-import {Dimensions} from 'react-native';
-
-export const getWidthWindow = () => {
-  const dim = Dimensions.get('window');
-  return dim.width;
-};
