+++ index.js
@@ -6,9 +6,9 @@ import {AppRegistry} from 'react-native';
 import {App} from './App';
 import {name as appName} from './app.json';
 import {setCustomTextInput, setCustomText} from 'react-native-global-props';
-import {customTextInputProps, customTextProps} from './App/consts/globalProps';
+import {commonTextStyle} from './App/commonStyles/textStyles';
 
-setCustomText(customTextProps);
-setCustomTextInput(customTextInputProps);
+setCustomText(commonTextStyle);
+setCustomTextInput(commonTextStyle);
 
 AppRegistry.registerComponent(appName, () => App);
